make pages into separate components, reorganize login page
