+++ src/lib/components/UserLoginUtil.svelte
@@ -1,0 +1,10 @@
+<script>
+	import { Checkbox } from 'flowbite-svelte';
+</script>
+
+<div class="container mx-auto flex w-full max-w-md items-start">
+	<Checkbox>Remember me</Checkbox>
+	<a href="/" class="ms-auto text-primary-700 hover:underline dark:text-primary-500">
+		Lost password?
+	</a>
+</div>
