+++ src/routes/+layout.svelte
@@ -5,8 +5,6 @@
 	import { Avatar, DarkMode, Navbar, NavBrand, NavHamburger, NavLi, NavUl } from 'flowbite-svelte';
 	import { MoonSolid, SunSolid } from 'flowbite-svelte-icons';
 	import '../app.css';
-
-	export let isLoggedIn = true;
 </script>
 
 <Navbar>
@@ -29,11 +27,7 @@
 		</FunctionalIcon>
 
 		<FunctionalIcon tooltip={'Registrieren oder einloggen'}>
-			<Avatar
-				rounded
-				class="apply-icon-style"
-				href={isLoggedIn ? '/userLand/userRegistration' : '/userLand/userLogin'}
-			/>
+			<Avatar rounded class="apply-icon-style" href={'/userLand/userLogin'} />
 		</FunctionalIcon>
 	</NavUl>
 </Navbar>
