+++ src/routes/userLand/userRegistration/+page.svelte
@@ -1,80 +1,5 @@
 <script>
-	import AbstractDropdownItem from '$lib/components/DataInput/AbstractDropdownItem.svelte';
-	import AbstractRegistrationForm from '$lib/components/DataInput/AbstractRegistrationForm.svelte';
-	import { Input } from 'flowbite-svelte';
-
-	const props = [
-		{
-			componentProps: {
-				label: 'Benutzername',
-				type: 'text',
-				placeholder: 'Wählen sie einen beliebigen Benutzernamen',
-				required: true,
-				about: 'Wählen sie einen beliebigen Benutzernamen'
-			},
-			itemComponent: Input
-		},
-		{
-			componentProps: {
-				label: 'E-Mail (optional)',
-				type: 'email',
-				placeholder: 'E-Mail',
-				about:
-					'Geben sie ihre E-mail Adresse an falls sie über aktuelle Forschungsergebnisse informiert werden möchten.'
-			},
-			itemComponent: Input
-		},
-		{
-			componentProps: {
-				label: 'Passwort',
-				type: 'password',
-				placeholder: 'Passwort',
-				required: true,
-				about:
-					'Wählen sie ein Passwort aus. Ihr login key wird aus diesem Passwort und ihrem Benutzernamen generiert.'
-			},
-			itemComponent: Input
-		},
-		{
-			componentProps: {
-				label: 'Passwort wiederholen',
-				type: 'password',
-				placeholder: 'Passwort wiederholen',
-				required: true,
-				about: 'Bestätigen sie ihr Passwort'
-			},
-			itemComponent: Input
-		},
-		{
-			componentProps: {
-				name: 'Benutzerrolle',
-				items: ['Benutzer/in', 'Admin', 'Forscher/in'],
-				about: 'Welche Rolle soll der neue Benutzer haben?',
-				selected: 'Benutzer/in'
-			},
-			itemComponent: AbstractDropdownItem
-		}
-	];
-
-	const buttons = [
-		{
-			label: 'Home',
-			href: '/'
-		},
-
-		{
-			label: 'Info',
-			href: '/info'
-		},
-		{
-			label: 'Weiter',
-			href: '/userLand/userDataInput'
-		}
-	];
-
-	const heading = 'Als neuer Benutzer registrieren';
+	import UserRegistration from '$lib/components/UserRegistration.svelte';
 </script>
 
-<div class="container m-1 mx-auto w-full max-w-md">
-	<AbstractRegistrationForm {heading} {props} {buttons} />
-</div>
+<UserRegistration />
